feat(config): enable formatters for lua and python
This commit updates the conform configuration to enable formatters for
Lua and Python. The formatters used are "stylua" for Lua and "black"
and "isort" for Python. Additionally, it removes commented-out
formatters for CSS and HTML. The format_on_save option is also
configured with a timeout of 1000ms and lsp_fallback set to true.